+++ pages/api/v1/sessions/index.ts
@@ -33,6 +33,7 @@ async function getHandler(
     email: user.email,
     created_at: user.created_at,
     updated_at: user.updated_at,
+    is_admin: user.is_admin,
   });
 }