+++ pages/api/v1/users/[username]/index.ts
@@ -41,6 +41,7 @@ async function getHandler(
     email: user.email,
     created_at: user.created_at,
     updated_at: user.updated_at,
+    is_admin: user.is_admin,
   });
 }
 
@@ -70,6 +71,7 @@ async function patchHandler(
     email: updatedUser.email,
     created_at: updatedUser.created_at,
     updated_at: updatedUser.updated_at,
+    is_admin: updatedUser.is_admin,
   });
 }