+++ pages/api/v1/users/index.ts
@@ -26,5 +26,6 @@ async function postHandler(
     email: newUser.email,
     created_at: newUser.created_at,
     updated_at: newUser.updated_at,
+    is_admin: newUser.is_admin,
   });
 }