+++ tests/integration/api/v1/users/[username]/patch.test.ts
@@ -126,6 +126,7 @@ describe("PATCH /api/v1/users/[username]", () => {
         password: responseBody.password,
         created_at: responseBody.created_at,
         updated_at: responseBody.updated_at,
+        is_admin: false,
       });
 
       expect(uuidVersion(responseBody.id)).toBe(4);
@@ -164,6 +165,7 @@ describe("PATCH /api/v1/users/[username]", () => {
         password: responseBody.password,
         created_at: responseBody.created_at,
         updated_at: responseBody.updated_at,
+        is_admin: false,
       });
 
       expect(uuidVersion(responseBody.id)).toBe(4);
@@ -204,6 +206,7 @@ describe("PATCH /api/v1/users/[username]", () => {
         password: responseBody.password,
         created_at: responseBody.created_at,
         updated_at: responseBody.updated_at,
+        is_admin: false,
       });
 
       expect(uuidVersion(responseBody.id)).toBe(4);