+++ types/index.ts
@@ -11,6 +11,7 @@ export interface User {
   created_at: Date;
   updated_at: Date;
   deleted_at: Date | null;
+  is_admin: boolean;
 }
 
 export interface UserPublic {
@@ -19,6 +20,7 @@ export interface UserPublic {
   email: string;
   created_at: Date;
   updated_at: Date;
+  is_admin: boolean;
 }
 
 export interface Session {